Move contentSub and contentSub2
There's a bug with displaying contentSub and contentSub2 on many
special pages. Moving them into the bodycontent fixes this.

Also, This change adds a configuration option to always use a
standard layout style, rather than using a standard or custom one
based on whether or not the content is styled.